editoast: adapt project to use model
