+++ editoast/src/main.rs
@@ -8,7 +8,6 @@ mod infra;
 mod infra_cache;
 mod map;
 mod models;
-mod projects;
 mod schema;
 mod tables;
 mod views;
