+++ editoast/src/views/projects.rs
@@ -1,14 +1,46 @@
 use crate::error::Result;
+use crate::models::Create;
+use crate::models::Delete;
 use crate::models::Document;
-use crate::projects::{Project, ProjectPatchForm};
-use crate::projects::{ProjectCreateForm, ProjectWithStudies};
+use crate::models::Project;
+use crate::models::ProjectWithStudies;
+use crate::models::Retrieve;
 use crate::views::pagination::{PaginatedResponse, PaginationQueryParam};
 use crate::DbPool;
 use actix_web::dev::HttpServiceFactory;
 use actix_web::web::{self, scope, Data, Json, Path, Query};
 use actix_web::{delete, get, patch, post, HttpResponse};
+use chrono::Utc;
+use derivative::Derivative;
+use editoast_derive::EditoastError;
+use image::ImageError;
+use serde::Deserialize;
 use serde::Serialize;
+use thiserror::Error;
 
+/// Returns `/projects` routes
+pub fn routes() -> impl HttpServiceFactory {
+    web::scope("/projects")
+        .service((create, list))
+        .service(scope("/{project}").service((get, delete, patch)))
+}
+
+#[derive(Debug, Error, EditoastError)]
+#[editoast_error(base_id = "project")]
+enum ProjectError {
+    /// Couldn't found the project with the given id
+    #[error("Project '{project_id}', could not be found")]
+    #[editoast_error(status = 404)]
+    NotFound { project_id: i64 },
+    // Couldn't found the project with the given id
+    #[error("Image document '{document_key}' not found")]
+    ImageNotFound { document_key: i64 },
+    // Couldn't found the project with the given id
+    #[error("The provided image is not valid : {0}")]
+    ImageError(ImageError),
+}
+
+/// Expand a project with its image url
 #[derive(Serialize, Debug, Clone)]
 struct ProjectWithImageUrl {
     image_url: Option<String>,
@@ -23,11 +55,50 @@ impl From<ProjectWithStudies> for ProjectWithImageUrl {
     }
 }
 
-/// Returns `/projects` routes
-pub fn routes() -> impl HttpServiceFactory {
-    web::scope("/projects")
-        .service((create, list))
-        .service(scope("/{project}").service((get, delete, patch)))
+/// Creation form for a project
+#[derive(Serialize, Deserialize, Derivative)]
+#[derivative(Default)]
+struct ProjectCreateForm {
+    pub name: String,
+    #[serde(default)]
+    pub description: String,
+    #[serde(default)]
+    pub objectives: String,
+    #[serde(default)]
+    pub funders: String,
+    #[serde(default)]
+    pub budget: i32,
+    pub image: Option<i64>,
+    #[serde(default)]
+    pub tags: Vec<String>,
+}
+
+impl From<ProjectCreateForm> for Project {
+    fn from(project: ProjectCreateForm) -> Self {
+        Project {
+            name: Some(project.name),
+            description: Some(project.description),
+            objectives: Some(project.objectives),
+            funders: Some(project.funders),
+            budget: Some(project.budget),
+            image: project.image.map(Some),
+            tags: Some(project.tags),
+            creation_date: Some(Utc::now().naive_utc()),
+            ..Default::default()
+        }
+    }
+}
+
+async fn check_image_content(db_pool: Data<DbPool>, document_key: i64) -> Result<()> {
+    let doc = match Document::retrieve(db_pool, document_key).await? {
+        Some(doc) => doc,
+        None => return Err(ProjectError::ImageNotFound { document_key }.into()),
+    };
+
+    if let Err(e) = image::load_from_memory(&doc.data.unwrap()) {
+        return Err(ProjectError::ImageError(e).into());
+    }
+    Ok(())
 }
 
 #[post("")]
@@ -35,10 +106,17 @@ async fn create(
     db_pool: Data<DbPool>,
     data: Json<ProjectCreateForm>,
 ) -> Result<Json<ProjectWithImageUrl>> {
-    let data = data.into_inner();
-    let project = Project::create(data, db_pool).await?;
+    let project: Project = data.into_inner().into();
+    if let Some(Some(image)) = project.image {
+        check_image_content(db_pool.clone(), image).await?;
+    }
+    let project = project.create(db_pool).await?;
+    let project_with_studies = ProjectWithStudies {
+        project,
+        studies: vec![],
+    };
 
-    Ok(Json(project.into()))
+    Ok(Json(project_with_studies.into()))
 }
 
 /// Return a list of projects
@@ -57,17 +135,52 @@ async fn list(
 /// Return a specific project
 #[get("")]
 async fn get(db_pool: Data<DbPool>, project: Path<i64>) -> Result<Json<ProjectWithImageUrl>> {
-    let project = project.into_inner();
-    let project = Project::retrieve(db_pool.clone(), project).await?;
-    Ok(Json(project.into()))
+    let project_id = project.into_inner();
+    let project = match Project::retrieve(db_pool.clone(), project_id).await? {
+        Some(project) => project,
+        None => return Err(ProjectError::NotFound { project_id }.into()),
+    };
+    let project_studies = project.with_studies(db_pool).await?;
+    Ok(Json(project_studies.into()))
 }
 
 /// Delete a project
 #[delete("")]
 async fn delete(project: Path<i64>, db_pool: Data<DbPool>) -> Result<HttpResponse> {
-    let project = project.into_inner();
-    Project::delete(project, db_pool).await?;
-    Ok(HttpResponse::NoContent().finish())
+    let project_id = project.into_inner();
+    if Project::delete(db_pool, project_id).await? {
+        Ok(HttpResponse::NoContent().finish())
+    } else {
+        Err(ProjectError::NotFound { project_id }.into())
+    }
+}
+
+/// Patch form for a project
+#[derive(Serialize, Deserialize)]
+struct ProjectPatchForm {
+    pub name: Option<String>,
+    pub description: Option<String>,
+    pub objectives: Option<String>,
+    pub funders: Option<String>,
+    pub budget: Option<i32>,
+    pub image: Option<i64>,
+    pub tags: Option<Vec<String>>,
+}
+
+impl ProjectPatchForm {
+    fn into_project(self, project_id: i64) -> Project {
+        Project {
+            id: Some(project_id),
+            name: self.name,
+            description: self.description,
+            objectives: self.objectives,
+            funders: self.funders,
+            budget: self.budget,
+            image: Some(self.image),
+            tags: self.tags,
+            ..Default::default()
+        }
+    }
 }
 
 #[patch("")]
@@ -76,14 +189,23 @@ async fn patch(
     project: Path<i64>,
     db_pool: Data<DbPool>,
 ) -> Result<Json<ProjectWithImageUrl>> {
-    let project = project.into_inner();
-    let project = Project::update(data.into_inner(), db_pool, project).await?;
-    Ok(Json(project.into()))
+    let data = data.into_inner();
+    let project_id = project.into_inner();
+    let project = data.into_project(project_id);
+    if let Some(Some(image)) = project.image {
+        check_image_content(db_pool.clone(), image).await?;
+    }
+    let project = match project.update(db_pool.clone()).await? {
+        Some(project) => project,
+        None => return Err(ProjectError::NotFound { project_id }.into()),
+    };
+    let project_studies = project.with_studies(db_pool).await?;
+    Ok(Json(project_studies.into()))
 }
 
 #[cfg(test)]
 mod test {
-    use crate::projects::Project;
+    use crate::models::Project;
     use crate::views::tests::create_test_service;
     use actix_http::Request;
     use actix_web::http::StatusCode;
